day6/oTree/requirements_base.txt
otree-core>=1.5.0b1
Django==1.11.2
django-jsonfield==1.0.1
python-Levenshtein==0.12.0
